refactor: server.py를 기능별 모듈로 분리
server.py(1200줄)를 4개 파일로 분리:
- data.py   (~107줄): 경로 상수, 데이터 I/O, 탐색 헬퍼
- ai.py     (~314줄): AI provider 호출, 시스템 프롬프트, 검증, 재시도
- os_utils.py (~113줄): 터미널·브라우저 탭 닫기 OS 유틸
- server.py (~703줄): HTTP Handler + 라우팅 + 진입점만 유지

ARCHITECTURE.md에 신규 파일 3개 항목 추가
